The packer's compressed float fused into an FMA on arm64 and wrote wrong bytes
The new compressed float path computed the quantized integer as one expression,
and Go permits fusing a multiply and an add into a single FMA unless a
conversion forces the intermediate rounding. The arm64 backend takes that
permission -- the line compiled to one FMADDS and rounded ONCE where the
runtimes round twice.

That made the DATA COMPILER architecture dependent: the same schema and the same
JSON packed on Apple Silicon and on x86-64 produced different bytes. Concretely,
compressed_float_value 0.005 over [0,10] at resolution 0.01 packed identical to
0.0 -- quantized 0, where every runtime writes 1.

The inner float32() conversion forces the rounding. serialize's STANDARD.md now
specifies the arithmetic normatively, since leaving it unstated is what let the
five implementations drift.

The existing test could not have caught this: its only compressed float is 2.5,
which normalizes to 0.25 and quantizes exactly, so it agrees under float32,
double and FMA alike -- which is why replacing the arithmetic with float64
survived the whole suite.

Companion fixes: mas-bandwidth/serialize#42 (C++ fused the same way at -O2),
serialize.c (was quantizing in double), serialize.go (same FMA).

+++ internal/pack/encode.go
@@ -237,7 +237,16 @@ func (e *Encoder) encodeScalar(w *bitWriter, f *ir.Field, val any, fpath string)
 				normalized = 1.0
 				e.warnf("%s: %v is above the declared float range [%v, %v] — CLAMPED (serialize_compressed_float's own write behaviour; fix the data)", fpath, fv, f.FMin, f.FMax)
 			}
-			quantized := uint64(math.Floor(float64(normalized*float32(maxIntegerValue) + 0.5)))
+			// The inner float32() conversion is LOAD BEARING, not decoration. Go
+			// permits fusing a multiply and an add into a single FMA unless an
+			// explicit conversion forces the intermediate rounding, and arm64
+			// takes that permission: without it this line compiles to one FMADDS
+			// and rounds ONCE where the runtimes round twice. That is not a
+			// rounding nicety — it changes the bytes. compressed_float_value
+			// 0.005 over [0, 10] resolution 0.01 quantizes to 0 fused and 1
+			// unfused, so the same source packed on arm64 and amd64 would emit
+			// different wire. Keep the conversion.
+			quantized := uint64(math.Floor(float64(float32(normalized*float32(maxIntegerValue)) + 0.5)))
 			w.writeBits(quantized, wireBits)
 			return nil
 		}
